# src/RATools.Api/openapi.v1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "RATools.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/document-placements": {
      "get": {
        "tags": [
          "DocumentPlacements"
        ],
        "parameters": [
          {
            "name": "applicationId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentPlacementDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentPlacementDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentPlacementDto"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "DocumentPlacements"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDocumentPlacementRequestBody"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDocumentPlacementRequestBody"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDocumentPlacementRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentPlacementDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentPlacementDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentPlacementDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/document-placements/{id}": {
      "delete": {
        "tags": [
          "DocumentPlacements"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/document-placements/{id}/section": {
      "put": {
        "tags": [
          "DocumentPlacements"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDocumentPlacementSectionRequestBody"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDocumentPlacementSectionRequestBody"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDocumentPlacementSectionRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentPlacementDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentPlacementDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentPlacementDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/document-placements/{id}/metadata": {
      "put": {
        "tags": [
          "DocumentPlacements"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDocumentPlacementMetadataRequestBody"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDocumentPlacementMetadataRequestBody"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDocumentPlacementMetadataRequestBody"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentPlacementDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentPlacementDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentPlacementDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/documents": {
      "get": {
        "tags": [
          "Documents"
        ],
        "parameters": [
          {
            "name": "applicationId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "sequenceNumber",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentDto"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentDto"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DocumentDto"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/documents/{id}": {
      "get": {
        "tags": [
          "Documents"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDto"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Documents"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        }
      }
    },
    "/api/documents/upload": {
      "post": {
        "tags": [
          "Documents"
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "File"
                ],
                "type": "object",
                "properties": {
                  "File": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "File": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/applications/{applicationId}/sequences/{sequenceNumber}/documents/upload": {
      "post": {
        "tags": [
          "Documents"
        ],
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "sequenceNumber",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "required": [
                  "CtdSection",
                  "File"
                ],
                "type": "object",
                "properties": {
                  "File": {
                    "type": "string",
                    "format": "binary"
                  },
                  "CtdSection": {
                    "maxLength": 128,
                    "minLength": 2,
                    "type": "string"
                  }
                }
              },
              "encoding": {
                "File": {
                  "style": "form"
                },
                "CtdSection": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/applications/{applicationId}/ectd-structure": {
      "get": {
        "tags": [
          "EctdStructure"
        ],
        "parameters": [
          {
            "name": "applicationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/EctdStructureDto"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EctdStructureDto"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/EctdStructureDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateDocumentPlacementRequestBody": {
        "required": [
          "applicationId",
          "ctdSection",
          "documentId",
          "operation",
          "sequenceNumber"
        ],
        "type": "object",
        "properties": {
          "documentId": {
            "type": "string",
            "format": "uuid"
          },
          "applicationId": {
            "type": "string",
            "format": "uuid"
          },
          "sequenceNumber": {
            "minLength": 1,
            "pattern": "^\\d{4}$",
            "type": "string"
          },
          "ctdSection": {
            "maxLength": 128,
            "minLength": 2,
            "type": "string"
          },
          "operation": {
            "maxLength": 32,
            "minLength": 2,
            "type": "string"
          },
          "title": {
            "maxLength": 512,
            "minLength": 0,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DocumentDto": {
        "required": [
          "createdUtc",
          "fileName",
          "fileSize",
          "id",
          "md5",
          "mediaType",
          "sha256",
          "storagePath"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "fileName": {
            "type": "string"
          },
          "mediaType": {
            "type": "string"
          },
          "fileSize": {
            "type": "integer",
            "format": "int64"
          },
          "sha256": {
            "type": "string"
          },
          "md5": {
            "type": "string"
          },
          "storagePath": {
            "type": "string"
          },
          "createdUtc": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "DocumentPlacementDto": {
        "required": [
          "applicationId",
          "createdUtc",
          "ctdSection",
          "documentId",
          "id",
          "lifecycleTargetPlacementId",
          "operation",
          "sequenceNumber",
          "title"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "documentId": {
            "type": "string",
            "format": "uuid"
          },
          "applicationId": {
            "type": "string",
            "format": "uuid"
          },
          "sequenceNumber": {
            "type": "string"
          },
          "ctdSection": {
            "type": "string"
          },
          "operation": {
            "type": "string"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "lifecycleTargetPlacementId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "createdUtc": {
            "type": "string",
            "format": "date-time"
          }
        },
        "additionalProperties": false
      },
      "EctdStructureDto": {
        "required": [
          "profileName",
          "region",
          "roots"
        ],
        "type": "object",
        "properties": {
          "profileName": {
            "type": "string"
          },
          "region": {
            "type": "string"
          },
          "roots": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EctdStructureNodeDto"
            }
          }
        },
        "additionalProperties": false
      },
      "EctdStructureNodeDto": {
        "required": [
          "children",
          "displayName",
          "elementName",
          "sectionPath",
          "sourceProfile"
        ],
        "type": "object",
        "properties": {
          "elementName": {
            "type": "string"
          },
          "sectionPath": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "sourceProfile": {
            "type": "string"
          },
          "children": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EctdStructureNodeDto"
            }
          }
        },
        "additionalProperties": false
      },
...
      "UpdateDocumentPlacementMetadataRequestBody": {
        "required": [
          "fileNamePrefix",
          "operation"
        ],
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "nullable": true
          },
          "operation": {
            "maxLength": 32,
            "minLength": 2,
            "type": "string"
          },
          "lifecycleTargetPlacementId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "fileNamePrefix": {
            "maxLength": 255,
            "minLength": 1,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "UpdateDocumentPlacementSectionRequestBody": {
        "required": [
          "ctdSection"
        ],
        "type": "object",
        "properties": {
          "ctdSection": {
            "maxLength": 128,
            "minLength": 2,
            "type": "string"
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}